pyproject.toml
[project]
name = "rag-paper-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index>=0.14.13",
    "llama-index-vector-stores-chroma>=0.5.5",
    "llama-parse>=0.6.54",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
